requirements.txt
Flask
Werkzeug
numpy
opencv-python
tensorflow
Pillow
math